Some progress

+++ src/mrtooley/core/storage_sqlite.py
@@ -8,14 +8,42 @@ import sqlite3
 import weakref
 from mrtooley.core.datatypes import Serializer
 from pathlib import Path
-from mrtooley.core.storage import StorageMapping, StorageBackend, StorageError
+from mrtooley.core.storage import StorageMapping, StorageBackend, StorageError, enable_deep_path_lookup
 from threading import Lock
-from typing import Optional
+from typing import Optional, Union
 
 
-class SqliteStorage(StorageMapping):
-    def __init__(self, connection, root_id: int):
+class SqliteMapping(StorageMapping):
+    def __init__(self, storage: "SqliteFile", root_id: Optional[int]):
         StorageMapping.__init__(self)
+        self._storage = storage
+        self._root_id = root_id
+
+    @enable_deep_path_lookup
+    def __getitem__(self, key: str):
+        print(f"getting value for {key}...")
+        if key == "root":
+            return 1
+
+        if key == "path":
+            return SqliteMapping(self._storage, 1)
+
+        if key == "sublevel" and self._root_id == 1:
+            return "subvalue"
+
+        if key == "newpath":
+            return SqliteMapping(self._storage, 2)
+
+
+        return "fallback value"
+
+    @enable_deep_path_lookup
+    def __setitem__(self, key: str, value):
+        if isinstance(value, (StorageMapping, dict)):
+            # Create a directory
+            print("created dict on", key)
+
+        print(f"setting value for {key} to {value}")
 
 
 class SqliteFile(StorageBackend):
@@ -25,7 +53,7 @@ class SqliteFile(StorageBackend):
         StorageBackend.__init__(self)
         self._lock = Lock()
 
-        self._connections = []
+        self._connections: dict[Union[id, None], SqliteMapping] = {}
 
         self._db_file = db_file
         missing = db_file is None or not db_file.is_file()
@@ -74,11 +102,30 @@ class SqliteFile(StorageBackend):
     def _read_path_index(self):
         pass
 
-    def get_storage_root(self) -> SqliteStorage:
+    def get_storage_root(self) -> SqliteMapping:
+        return SqliteMapping(self, None)
+
+    def flush(self):
+        pass
+
+    def unload(self):
         pass
 
 
 db_path = Path("/tmp/testdb.sqlite")
-db_path.unlink()
+if db_path.is_file():
+    db_path.unlink()
 db = SqliteFile(db_path)
 
+rootstorage = db.get_storage_root()
+
+rootstorage["root"] = 1
+print(rootstorage["root"])
+print(rootstorage["path/sublevel"])
+
+
+rootstorage["path/sublevel2"] = 2
+
+
+rootstorage["newpath/sublevel3"] = 3
+print(rootstorage["newpath"])
